Added paymentStatusUpdate and completionStatusUpdate after successfully order placed for tracking

+++ src/main/java/com/kalsym/order/service/model/Order.java
@@ -40,10 +40,10 @@ public class Order implements Serializable {
     private String id;
 
     private String storeId;
-    private Float subTotal;
-    private Float serviceCharges;
+    private Double subTotal;
+    private Double serviceCharges;
     private Double deliveryCharges;
-    private Float total;
+    private Double total;
     @Enumerated(EnumType.STRING)
     private OrderStatus completionStatus;
     @Enumerated(EnumType.STRING)
